PlaceVerificationInfoView UI 완료
